+++ CHANGELOG.md
@@ -6,8 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ---
 
+## [1.1.1-beta.8] -2026-06-11
+
+### Add New Language Support
+
+- Add more 5 language support.
+- Such as `kannada`, `marathi`, `odia`, `tamil`, and `telugu`.
+- Support for more languages will be added in the future.
+- Update previous language support.
+- The previous language add morethen 200+ slangs.
+- The new language add morethen 200+ slangs.
+
+---
+
 ## [1.0.0-beta.6] - 2024-10-13
+
 ### Fixed an issue
+
 - Added new functionality to improve the use of slangs.
 - Fixed an issue with slangs that caused the use of slangs.
 - Minor bug fixes and performance improvements.
@@ -15,29 +30,40 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ---
 
 ## [1.0.0-beta.5] - 2024-10-13
+
 ### add languages
+
 - add new module `languages`
 
 ---
 
 ## [1.0.0-beta.4] - 2024-10-12
+
 ### Changed
+
 - Updated API responses for better consistency.
+
 ---
 
 ## [1.0.0-beta.3] - 2024-10-12
+
 ### changed
+
 - Introduced initial support for slangs.
 
 ---
 
 ## [1.0.0-beta.1] - 2024-10-12
+
 ### Changed
+
 - Refactored code for slangs to improve maintainability.
 
 ---
 
 ## [1.0.0-beta.0] - 2024-10-11
+
 ### Initial Release
+
 - This is the first beta release with the core functionality of slangs.
 - Includes basic setup and slangs.